Revert "* Trap packet if ecmp index is set to -1 or there is no forwarding meta data"
This reverts commit f9f3815ac0b2c2fd91f2953a62b1cc89e204a249.
By mistake committed directly to master, will commit via pull request

+++ dp-core/vr_nexthop.c
@@ -347,7 +347,7 @@ nh_composite_ecmp(unsigned short vrf, struct vr_packet *pkt,
     if (stats)
         stats->vrf_ecmp_composites++;
 
-    if (!fmd || (fmd->fmd_ecmp_nh_index >= (short)nh->nh_component_cnt))
+    if (!fmd || (uint8_t)fmd->fmd_ecmp_nh_index >= nh->nh_component_cnt)
         goto drop;
 
     if (fmd->fmd_ecmp_nh_index >= 0)
